+++ src/main/java/app/visa/repository/NationaliteRepository.java
@@ -3,5 +3,5 @@ package app.visa.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import app.visa.entity.Nationalite;
 
-public interface NationaliteRepository extends JpaRepository<Nationalite, Long> {
+public interface NationaliteRepository extends JpaRepository<Nationalite, Integer> {
 }
