+++ src/main/java/app/visa/repository/VisaTransformableRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import app.visa.entity.VisaTransformable;
 
-public interface VisaTransformableRepository extends JpaRepository<VisaTransformable, Long> {
+public interface VisaTransformableRepository extends JpaRepository<VisaTransformable, Integer> {
 }
